requirements.txt
discord.py
emoji
colorama
motor
python-dateutil
aiohttp
aiofiles